Split out the text support from the extensions
